change added

+++ cart-service/src/main/java/com/birlasoft/api/controller/CartController.java
@@ -1,5 +1,7 @@
 package com.birlasoft.api.controller;
 
+import java.util.Optional;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,8 +20,6 @@ import com.birlasoft.api.commons.Request;
 import com.birlasoft.api.entity.Cart;
 import com.birlasoft.api.service.CartService;
 
-
-
 @RestController
 //@RequestMapping("/cart")
 
@@ -28,41 +28,58 @@ public class CartController {
 	@Autowired
 	private CartService service;
 
-	static Logger log=LoggerFactory.getLogger(CartController.class);
-	
+	static Logger log = LoggerFactory.getLogger(CartController.class);
+
 	@PostMapping("/product")
 	public ResponseEntity addProductInCart(@RequestBody Request request) {
-		
-		log.info("addProductInCart  method is running");
-		log.info("Request "+request);
 
-		
+		log.info("addProductInCart  method is running");
+		log.info("Request " + request);
+
 		try {
-			Cart saveCart=service.addInCart(request);
-			if(request.getProductId()!=0 && saveCart != null) {
-				return new ResponseEntity<Cart>(saveCart,HttpStatus.OK);	 
-			}
-			else
-				return new ResponseEntity<String>("Product not Added in Cart Successsfully",HttpStatus.OK);
-		
+			Cart saveCart = service.addInCart(request);
+			if (request.getProductId() != 0 && saveCart != null) {
+				return new ResponseEntity<Cart>(saveCart, HttpStatus.CREATED);
+			} else
+				return new ResponseEntity<String>("Product not Added in Cart Successsfully", HttpStatus.OK);
+
 		} catch (Exception e) {
-			return new ResponseEntity<String>(e.getMessage(),HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<String>(e.getMessage(), HttpStatus.BAD_REQUEST);
 		}
 
-		
 	}
 
 	@DeleteMapping("/product/{id}")
-	public void removeProductById(@PathVariable("id") int cartId) {
+	public ResponseEntity removeProductById(@PathVariable("id") int cartId) {
 		log.info("removeProductById method");
-		service.removeFromCart(cartId);
+		try {
+			if (cartId != 0) {
+				service.removeFromCart(cartId);
+				return new ResponseEntity<Void>(HttpStatus.OK);
+			} else
+				return new ResponseEntity<String>("Product not Added in Cart Successsfully", HttpStatus.OK);
+		} catch (Exception e) {
+			return new ResponseEntity<String>(e.getMessage(), HttpStatus.BAD_REQUEST);
+		}
+
 	}
 
 	@GetMapping("/product/{userName}")
 	public ResponseEntity getProductByUserName(@PathVariable("userName") String userName) {
 
-		log.info("getProductByUserName method with User Name: "+ userName);
-		return new ResponseEntity<>(service.getByUserName(userName), HttpStatus.OK);
+		log.info("getProductByUserName method with User Name: " + userName);
+		try {
+			Optional<Cart> response = service.getByUserName(userName);
+			if (userName != null && response != null) {
+				return new ResponseEntity<Optional<Cart>>(response, HttpStatus.OK);
+			} else {
+				return new ResponseEntity<String>("Cart is empty or user is not registered", HttpStatus.OK);
+			}
+		}
+
+		catch (Exception e) {
+			return new ResponseEntity<String>(e.getMessage(), HttpStatus.BAD_REQUEST);
+		}
 	}
 
 }
